requirements.txt
requests
pandas
python-telegram-bot==20.3
